pyproject.toml
[project]
name = "bank-churn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
   "evidently>=0.7.9",
   "flask>=3.1.1",
   "google-cloud-storage>=3.2.0",
   "gunicorn>=23.0.0",
   "imbalanced-learn==0.13.0",
   "imblearn>=0.0",
   "ipykernel>=6.29.5",
   "jupyter>=1.1.1",
   "kaggle>=1.7.4.5",
   "lightgbm>=4.6.0",
   "make>=0.1.6.post2",
   "mlflow>=3.1.1",
   "pandas>=2.3.0",
   "pip>=25.1.1",
   "prefect>=3.4.7",
   "prefect-gcp>=0.6.8",
   "psycopg2-binary>=2.9.10",
   "pydantic>=2.11.7",
   "python-dotenv>=1.1.1",
   "scikit-learn==1.5.2",
   "seaborn>=0.13.2",
   "xgboost>=3.0.2",
]

[dependency-groups]
dev = [
   "pre-commit>=4.2.0",
   "pytest>=8.4.1",
   "pytest-cov>=6.2.1",
   "ruff>=0.12.1",
]

[tool.ruff]
# Set the maximum line length to 88 (same as Black)
line-length = 88
target-version = "py311"

# Enable specific rule sets
lint.select = [
   "E",    # pycodestyle errors
   "W",    # pycodestyle warnings
   "F",    # Pyflakes
   "I",    # isort
   "N",    # pep8-naming
   "UP",   # pyupgrade
   "B",    # flake8-bugbear
   "C4",   # flake8-comprehensions
]

# Ignore specific rules that are too strict for this project
lint.ignore = [
   "E501",   # Line too long (handled by formatter)
   "B008",   # Do not perform function calls in argument defaults
   "N806",   # Allow ML variable naming conventions (X_train, X_test, etc.)
   "N803",   # Allow ML arg naming conventions
   "E402",   # Module level import allowed at top of file
]

# Exclude directories
exclude = [
   ".git",
   ".venv",
   "__pycache__",
   "build",
   "dist",
]

[tool.ruff.lint.isort]
# Configure import sorting
known-first-party = ["mlops_churn"]
